Update pom.xml and clean up QCServer.java
In pom.xml, the version of QCServer is updated from 1.2.6 to 1.2.7, log4j dependency is removed and replaced by slf4j. Several other dependencies versions are updated. In QCServer.java, unused imports are removed. These updates are expected to enhance security and performance of the code, while the removal of unused imports in QCServer.java results in cleaner code.